refactor: moving logic from controller to service
