+++ src/api/user/user.controller.ts
@@ -1,36 +1,16 @@
-import { RequestContext } from '@mikro-orm/core';
-import { User } from './user.entity.js';
-import ApiException from '../../core/ApiException.js';
-import * as authService from '../auth/auth.service.js';
-import { NextFunction, Request, Response } from 'express';
+import type { NextFunction, Request, Response } from 'express';
+import * as userService from './user.service.js';
+import { CreateUserApiResponse } from './user.types.js';
 
-export const createUser = async (req: Request, res: Response, next: NextFunction): Promise<void> => {
-    const em = RequestContext.getEntityManager();
-    const repository = em?.getRepository(User);
-    const exists = await repository?.exists(req.body.email);
-
-    if (exists) {
-        const error = new ApiException({
-            message: 'A user with that email already exists',
-            status: 409,
-        });
-        return next(error);
+export const createUser = async (
+    req: Request,
+    res: Response<CreateUserApiResponse>,
+    next: NextFunction,
+): Promise<void> => {
+    try {
+        const user = await userService.createUser(req.body.email, req.body.password);
+        res.send(user);
+    } catch (e) {
+        next(e);
     }
-
-    const hashedPassword = await authService.generateBcryptHash(req.body.password);
-    const newUser = new User(req.body.email, hashedPassword);
-
-    await em?.persist(newUser).flush();
-
-    const accessToken = authService.createUserAccessToken(newUser.id);
-    const refreshTokenEntity = authService.createRefreshTokenEntity(newUser.id);
-
-    await em?.persist(refreshTokenEntity).flush();
-
-    res.json({
-        userId: newUser.id,
-        email: newUser.email,
-        accessToken,
-        refreshToken: refreshTokenEntity.token,
-    });
 };
