+++ src/api/user/user.types.ts
@@ -1,0 +1,3 @@
+import { User } from './user.entity.js';
+
+export type CreateUserApiResponse = { id: User['id']; email: User['email']; accessToken: string; refreshToken: string };
